requirements.txt
fastapi
uvicorn
pyjwt[crypto]
cryptography
pytest
pytest-cov
httpx
ruff
